==> 01-Habitat-use.Rmd <==
# Difference in distribution and habitat use between _H. curtus_ and _H. shcistosus_

We investigated the study species used similar areas for foraging and other activities using survey data gathered from fishermen. We tested differences in depth use and overlap in space.

## Calculating abundance of sea snakes per fishing trip

```{r echo = TRUE}

source(file = "Functions/setup.R")

snakes_den = snakes%>%
  filter(Species == "Hydrophis schistosus" | Species == "Hydrophis curtus",
         Fishing.Location != "", 
         !is.na(Depth.Caught..m.))%>%
  dplyr::select(Species, Fishing.Location, Depth.Caught..m., Date, Boat.Name.Owner, # selecting relevant variables
         No..of.Hauls, Average.Haul.Duration..Hours., Gear.Type)%>%
  group_by(Gear.Type)%>% # completing missing effort data
  mutate(n.hauls = ifelse(is.na(No..of.Hauls), 
                         median(No..of.Hauls, na.rm = T), 
                         No..of.Hauls),
         haul.time = ifelse(is.na(Average.Haul.Duration..Hours.), 
                         median(Average.Haul.Duration..Hours., na.rm = T),
                         Average.Haul.Duration..Hours.))%>%
  group_by(Date, Boat.Name.Owner)%>%
  summarise(HC = sum(Species == "Hydrophis curtus"), # calculating abundance per trip
         HS = sum(Species == "Hydrophis schistosus"),
         effort = last(n.hauls*haul.time), # calculating fishing effort per trip
         Fishing.Location = last(Fishing.Location),
         Depth.Caught..m. = last(Depth.Caught..m.))%>%
  gather(c("HC", "HS"), value = n, key = "Species")%>%
  mutate(CPUE = n/effort)%>% # calculating catch per unit effort of each specie per trip
  drop_na() # removing missing values

```

## Sample size

The number of trips and total haul hours sampled by gear.

```{r}

snakes%>%
  filter(Species == "Hydrophis schistosus" | Species == "Hydrophis curtus",
         Fishing.Location != "", 
         !is.na(Depth.Caught..m.))%>%
  dplyr::select(Species, Fishing.Location, Depth.Caught..m., Date, Boat.Name.Owner, # selecting relavant variables
         No..of.Hauls, Average.Haul.Duration..Hours., Gear.Type)%>%
  group_by(Gear.Type)%>% # filling in missing effort data
  mutate(n.hauls = ifelse(is.na(No..of.Hauls),
                         median(No..of.Hauls, na.rm = T), No..of.Hauls),
         haul.time = ifelse(is.na(Average.Haul.Duration..Hours.), 
                         median(Average.Haul.Duration..Hours., na.rm = T),
                         Average.Haul.Duration..Hours.))%>%
  group_by(Gear.Type, Date, Boat.Name.Owner)%>%
  summarise(effort = last(n.hauls*haul.time))%>% # calculating effort per trip
  group_by(Gear.Type)%>%
  summarise(n_trips = n(),
            effort = sum(effort))%>%
  drop_na()

```

```{r}
snakes%>%
  filter(Species == "Hydrophis schistosus" | Species == "Hydrophis curtus")%>%
  summarise(HS = sum(Species == "Hydrophis schistosus"),
            HC = sum(Species == "Hydrophis curtus"))
  
```

## Geocoding sea snake encounters

In order to map sea snake distribution and study depth use we used location and depth data gathered from fisher surveys to map each fishing trip.

The protocol for geocoding fishing trips is as follows:

- Nearest landmarks are geocoded from google maps API 
- Latitude is extracted from landmark geocode
- Match landmark latitude and depth from survey to GEBCO data
- Extract final longitude from GEBCO data

The protocol is encoded in `Function/geocode.R`

```{r}

source("./Functions/geocode.R")

snakes_den <- geocode.trips(snakes_den, fl = 'Fishing.Location', dep = 'Depth.Caught..m.')

# removing duplicates introduced by fuzzy joining

snakes_den <- distinct(snakes_den, Date, Boat.Name.Owner, Species, Fishing.Location, Depth.Caught..m., .keep_all = T)

# saving output

write.csv(snakes_den, "./Data/snakes-density.csv")

```

**Note:** `fuzzyjoin::fuzzy_inner_join()` introduces many duplicates, I have not as yet been able to fix the issue directly. Currently, removing duplicates from output using `dplyr::distinct()`.

## Mean CPUE by species

Each fishing trip is assigned to a cell in the sampling grid. Total CPUE of each species in each grid cell is then calculated.

```{r}
# loading libraries

library(raster)

# raster with study site extent

ext <- raster("./Data/sampling_extent.tif")

# importing functions

source("./Functions/intensity extract.R")
source("./Functions/raster to df.R")

## Calculating sea snake CPUE

den <- snakes_den%>%
  group_by(Species)%>%
  nest()%>%
  mutate(m = map(data, ~map.extract(df = ., var = "n", func = 'sum')),
         mdf = map(m, map.df))%>%
  dplyr::select(mdf)%>%
  unnest()%>%
  spread(Species, layer)

# Mean, sd CPUE

den%>%
  skimr::skim(HC, HS)%>%
  skimr::yank("numeric")%>%
  dplyr::select(-hist)

```

## Area used by species

The area (`km^2`) occupied by a species calculated as the number of grid cells that a species was found in multiplied the area of each cell (10 sq. km.).

```{r}

snakes_den%>%
  group_by(Species)%>%
  nest()%>%
  mutate(m = map(data, ~map.extract(df = ., var = "CPUE", func = 'sum')),
         mdf = map(m, map.df))%>%
  dplyr::select(mdf)%>%
  summarise(spatial_extent = map(mdf, ~sum(.$layer>0)))%>%
  unnest()%>%
  mutate(spatial_extent = spatial_extent*10)

```

## Mapping sea snake distributions

```{r fig.height=8, fig.width=4}

## importing required libraries

library(marmap)
library(ggmap)

## getting deoth data downloaded from GEBCO

depth <- readGEBCO.bathy("./Data/gebco_2020_n16.7_s15.5_w72.0_e73.9.nc")

depth = fortify.bathy(depth)%>% # converting to usable data frame
  dplyr::rename(lon = x, lat = y, depth = z)%>%
  filter(depth < 1)


## Getting base map
map_center = c(lon = mean(snakes_den$lon), lat = mean(snakes_den$lat))

base <- get_googlemap(center = c(lon = mean(snakes_den$lon), lat = mean(snakes_den$lat)), 
                      zoom = 9, scale = 4, maptype = "terrain")##Getting basemap

## importing plotting function

source("./Functions/intensity_plot.R")

## making plot

snake_plot <- snakes_den%>%
  mutate(Species = ifelse(Species == "HC", "Hydrophis curtus", "Hydrophis schistosus"))%>%
  group_by(Species)%>%
  nest()%>%
  mutate(m = map(data, ~map.extract(df = ., var = "CPUE", func = 'sum')), # rasterising CPUE
         mdf = map(m, map.df), # Converting raster to df for plotting
         plot = map2(mdf, Species, ~int.plot(mdf = ., Species, name = "CPUE", title.face = "italic")) #plotting raster on base map
         )%>%
  gridExtra::grid.arrange(grobs = .$plot, ncol = 1)

## saving high res figure

ggsave(snake_plot, filename = "./Figures and Tables/figure1.tiff", width = 8, height = 16, dpi = 300)


```

**Figure 1**: Distribution and depth use of sea snake in the near shore waters of Sindhudurg, Maharashtra based on fisheries dependent data.

## Summarising depth use by species

```{r}
# Calculating mean depth in cell

dep <- depth%>%
  nest()%>%
  mutate(m = map(data, ~map.extract(df = ., var = "depth", func = mean)),
         mdf = map(m, map.df))%>%
  dplyr::select(mdf)%>%
  unnest()%>%
  rename(mean.depth = layer)

#Calculating relative abaundance of sea snakes in each cell

den<- den%>%
  inner_join(dep, by = c("x", "y"))%>%
  mutate(rel.prop = HC/(HC+HS))

# Relative proportion of H. curtus with depth

den%>%
  gather(c("HS", "HC"), key = sp, value = CPUE)%>%
  filter(CPUE>0)%>%
  group_by(sp)%>%
  skimr::skim(mean.depth)%>%
  skimr::yank("numeric")%>%
  dplyr::select(-hist)


```

## Spatial segregation by species

We caluclated the spatial overlap between _H. curtus_ and _H. schistosus_ as the area (number of cells*10 sq. km.) in which both species occured together divided the by the total area in which atleast one of the species occured.

```{r}
den%>%
  summarise(total.extent = sum(HC>0 | HS>0)*10,
            overlap = sum(HC > 0 & HS > 0)*10,
            relative_overlap = overlap/total.extent)

```


## Modelling depth use by species

```{r }
# t-test for cell depth

den%>%
  gather(c("HS", "HC"), key = sp, value = CPUE)%>%
  filter(CPUE>0)%>%
  dplyr::select(sp, mean.depth)%>%
  nest()%>%
  mutate(mod = map(data, ~t.test(mean.depth ~ sp, data = .)),
         sumry = map(mod, broom::tidy),
         d = map(data, ~lsr::cohensD(mean.depth ~ sp, data = .)))%>%
  dplyr::select(sumry, d)%>%
  unnest()%>%
  dplyr::select(estimate:parameter, d)

# testing effect of cell depth on relative abundance

den%>%
  gather(c("HS", "HC"), key = sp, value = CPUE)%>%
  mutate(rel.prop = ifelse(sp == "HS", 1 - rel.prop, rel.prop),
         mean.depth = -mean.depth)%>%
  dplyr::select(sp, rel.prop, mean.depth)%>%
  group_by(sp)%>%
  nest()%>%
  mutate(prop_dep = map(data, ~lm(rel.prop ~ mean.depth, data = .)),
         sumry = map(prop_dep, broom::tidy),
         stat = map(prop_dep, broom::glance))%>%
  dplyr::select(sumry, stat)%>%
  unnest()%>%
  dplyr::select(sp:p.value, adj.r.squared)
  

```

_H. curtus_ was caught in deeper habitats than _H. schistosus_.
