==> 02-Diet.rmd <==

# DIfference in diet between *H. curtus* and *H. schsitosus*

## No. of samples collected by sea snake species

```{r}
source("./Functions/setup.R")

# importing gut content data

gc = read.csv("./Data/Sea_snakes_gut_content_2018-19.csv")%>%
  filter(Source == "Gut content") # removing specimens collected from fisheries landings
```


```{r}
snakes%>%
  filter(Species == "Hydrophis curtus" | Species == "Hydrophis schistosus")%>%
  group_by(Species)%>%
  summarise(N = n(), # total number of individuals sampled
            n = sum((Gut.Content) == "Yes"), # number of individuals with gut content
            n_male = sum((Gut.Content) == "Yes" & Sex == "Male"), # males with gut content
            n_female = sum((Gut.Content) == "Yes" & Sex == "Female"), # females with content
            prop.gc = n/N) # propotion of individuals with gut content

```

A very low proportion of individuals had gut content present at the time of sampling. Sampling across species and sexes may be adequate for comparison.

## Total number of prey specimens collected

```{r}
gc%>%
  group_by(Snake.Species)%>%
  summarise(N = length(unique(GC.Code)))
```


## Proportion of unidentified specimens

```{r}

# calculating richness

gc%>%
  summarise(N_sp = length(unique(Prey.Species)), # number of prey species
            N_fam = length(unique(Prey.Family)), # number of prey families
            # proportion of unidentitified specimens
            unid_sp = sum(Prey.Species == "Unidentified" | Prey.Species == "", na.rm = T)*100/n(), ## species
            unid_fam = sum(Prey.Family == "Unidentified" | Prey.Family == "", na.rm = T)*100/n())%>% ## family
  # creating clean table
  gather()%>%
  mutate(Metric = ifelse(substr(key, 1, 1) == "N", "Richness", "% Unidentified"),
         Unit = ifelse(grepl(key, pattern = "sp", fixed = T), "Prey Species", "Prey Family"))%>%
  dplyr::select(Unit, Metric, value)%>%
  spread(Metric, value)
```

A large portion of gut content specimens were unidentifiable. This is an unavoidable consquence of VGCA.

## No. of snakes with mode than one prey specimen

```{r}
gc%>%
  filter(Field.Code != "")%>%
  count(Field.Code,name = "Prey")%>%
  count(Prey > 1)
```

## Prey preference

We modified and used the the Index of relative importance according to Pinkas et al. 1971 to determine the difference in prey preference between _H. curtus_ and _H. schistosus_

```{r}

# calulating index of relative importance (Pinkas et al. 1971)

IRI <- gc%>%
  filter(Prey.Family != "")%>%
  filter(Snake.Species == "Hydrophis schistosus" | Snake.Species == "Hydrophis curtus")%>%
  group_by(Snake.Species)%>%
  mutate(Fr = length(unique(Field.Code)), # total number of snakes sampled
         W = sum(Weight..g., na.rm = T), # total weight of prey sampled
         N = n())%>% # total number of prey sampled
  group_by(Snake.Species, Prey.Family)%>%
  summarise(f = length(unique(Field.Code)), # number snakes in which prey family occured
            Fr = last(Fr),
            w = sum(Weight..g., na.rm = T), # weight of prey family
            W = last(W),
            n = n(), # number of individuals of prey family
            N = last(N))%>%
  group_by(Snake.Species, Prey.Family)%>%
  # caluclating percentages
  summarise(per.F = f*100/Fr,
            per.W = w*100/W,
            per.N = n*100/N,
            # caluclating IRI
            IRI = (per.N + per.W)*per.F)%>%
  ungroup()%>%
  complete(Prey.Family, nesting(Snake.Species), fill = list(0))%>%
  replace_na(replace = list(IRI = 0, per.F = 0, per.W = 0, per.N = 0))%>%
  arrange(Snake.Species, desc(IRI))%>%
  mutate(rank = 1:n())

# clean table

IRI%>%
  filter(IRI > 0)

```

```{r fig.height=6, fig.width=8}

# plotting prey preference

IRI%>%
  ggplot(aes(reorder(Prey.Family, IRI), IRI, fill = Snake.Species))+
  geom_col(col = "black", position = position_dodge(preserve = "single"))+
  scale_y_sqrt(name = "IRI (sq.rt.)")+
  labs(x = "Prey Family")+
  scale_fill_brewer(palette = "Greys", name = "Species")+
  theme(axis.text.x = element_text(hjust = 1, angle = 60),
        legend.text = element_text(face = "italic"))

# saving high res plot

ggsave(last_plot(), filename = "./Figures and Tables/figure2.tiff", width = 8, height = 6, dpi = 300)

```

**Figure 2**: Index of relative importance (IRI, square root transformed) (Pinkas et al., 1970) of prey families in sea snake diet.

_H. schistosus_ showed preference for the Tetraodontid fish, while _H. curtus_ did not exhibit any prey preference.

## Range of prey found in _H. curtus_ and _H. shistosus_ gut

We compared the dietary breadth between snakes species using the richness and diversity of prey families.

```{r}

# Creating data matrix for prey community analysis analysis

fam_spade = gc%>%
  filter(Prey.Family != "Unidentified", Prey.Family !="", Field.Code != "")%>% # removing unidentified specimens
  group_by(Snake.Species, Prey.Family)%>%
  summarise(n = n())%>% # total number of specimens per family in each snake species
  spread(Snake.Species, n, fill = 0)%>%
  column_to_rownames("Prey.Family")


```

### Prey richness

```{r}

gc%>%
  filter(Snake.Species == "Hydrophis schistosus" | Snake.Species == "Hydrophis curtus",
         Prey.Family != "Unidentified", Prey.Family != "")%>%
  group_by(Snake.Species)%>%
  summarise(Prey.Species = length(unique(Prey.Species)) - 1, # removing unidentified
            Prey.Families = length(unique(Prey.Family)))

```

### Prey family diversity

```{r}

# loading libraries

library(SpadeR) ## for community indices

# calculating shanon diversity

HC_div = as.data.frame(Diversity(fam_spade[,1])$Shannon_diversity)%>%
  rownames_to_column("Estimator")%>%
  mutate(Species = "Hydrophis curtus")

HS_div = as.data.frame(Diversity(fam_spade[,2])$Shannon_diversity)%>%
  rownames_to_column("Estimator")%>%
  mutate(Species = "Hydrophis schistosus")

HC_div%>%
  full_join(HS_div)%>%
  dplyr::select(Species, everything())%>%
  arrange(Estimator)
```

While _H. schistosus_ fed on greater number of prey families, _H. curtus_ showed greater evenness in prey preference.

Statistical comparison is difficult as each individual snake fed on only one type of specimen at time.

## Diet overlap between _H. curtus_ and _H. shistosus_

Number of overlapping prey families:

```{r}

gc%>%
  filter(Snake.Species == "Hydrophis schistosus" | Snake.Species == "Hydrophis curtus",
         Prey.Family != "Unidentified", Prey.Family != "")%>%
  group_by(Prey.Family)%>%
  summarise(N_pred = length(unique(Snake.Species)))%>%
  ungroup()%>%
  summarise(Overlap = sum(N_pred>1))%>%
  as.numeric()

```

Shoener's (1971) resource overlap index (D):

```{r}

gc%>%
  filter(Prey.Species != "Unidentified")%>%
  group_by(Snake.Species, Prey.Family)%>%
  summarise(n = n())%>% # number of speciemens per prey family in each snake species
  group_by(Snake.Species)%>%
  mutate(N = sum(n))%>% # total number of specimens collected from each snake species
  group_by(Snake.Species, Prey.Family)%>%
  summarise(p = n/N)%>% # relative proportion of each prey family
  spread(Snake.Species, p, fill = 0)%>%
  rename("HC" = 'Hydrophis curtus', "HS" = 'Hydrophis schistosus')%>%
  mutate(P = abs(HC - HS))%>% # difference in relactive proportion prey items 
  summarise(D = 1 - 0.5*sum(P))%>% # Schoener's resource overlap index
  as.numeric()

```

Morista - Horn overlap:

```{r}
prey_sim <- SimilarityPair(X = fam_spade, datatype = "abundance")

prey_sim$Empirical_relative
```

There was limited overlap in prey between _H. schistosus_ and _H. curtus_

### Testing segregation in prey between _H. curtus_ and _H. schistosus_

```{r}

# loading libraries

library(vegan)

# formating data for vegan

fam_simboo <- gc%>%
  filter(Prey.Family != "Unidentified", Prey.Family !="", Field.Code != "")%>%
  group_by(Snake.Species, Field.Code, Prey.Family)%>%
  summarise(n = n())%>%
  spread(Prey.Family, n, fill = 0)%>%
  ungroup()

# PERMANOVA to compare composition

set.seed(2)

permanova_sp <- adonis2(fam_simboo[,3:length(fam_simboo)] ~ fam_simboo$Snake.Species,
                     data = fam_simboo[,1])

## clean table

broom::tidy(permanova_sp)
```

Prey composition was significantly different between the two snake species.

## Difference in size selectivity of prey between _H. curtus_ and _H. schistosus_

```{r}

# summary

gc%>%
  filter(Snake.Species == "Hydrophis schistosus" | Snake.Species == "Hydrophis curtus",
         Condition < 3)%>% # removing very digested specimens 
  group_by(Snake.Species)%>%
  skimr::skim(Maximum.Body.Girth..cm.)%>%
  skimr::yank("numeric")%>%
  dplyr::select(Snake.Species:p100)


# testing diff max prey width by species

gc%>%
  filter(Snake.Species == "Hydrophis schistosus" | Snake.Species == "Hydrophis curtus",
         Condition < 3, # removing digested specimens 
         !is.na(Maximum.Body.Girth..cm.))%>% # removing unrecorded data
  dplyr::select(Snake.Species, Maximum.Body.Girth..cm.)%>%
  droplevels()%>%
  nest()%>%
  mutate(test = map(data, ~t.test(Maximum.Body.Girth..cm. ~ Snake.Species, data = .)), # t test
         sumry = map(test, broom::tidy),
         d = map(data, ~lsr::cohensD(Maximum.Body.Girth..cm. ~ Snake.Species, data = .)))%>% # effect size
  dplyr::select(sumry, d)%>%
  unnest()%>%
  dplyr::select(estimate1:parameter, d)%>%
  rename(`Hydrophis curtus` = estimate1,
         `Hydrophis schistosus` = estimate2)
```

_H. schistosus_ fed on larger prey than _H. curtus_
