==> 05-Fisheries_spatial-overlap.rmd <==
# Does fishing intensity affect the distribution of *H. curtus* and *H. schistosus*?

```{r}
# loading required libraries

library(sp)
library(raster)

# importing data

dep_int <- read.csv("./Data/Fishing intensity_dep.csv") # geocoded fishing effort

snakes_den <- read.csv("./Data/snakes-density.csv") # sea snake occurence in bycatch

ext <- raster::raster("./Data/sampling_extent.tif") # sampling grid

# importing required scripts

source("./Functions/intensity extract.R")
source("./Functions/raster to df.R")

# Calculating fishing intensity

fi <- dep_int%>%
  group_by(Gear.Type)%>%
  nest()%>%
  mutate(m = map(data, ~map.extract(df = ., var = "effort", func = 'sum')),
         mdf = map(m, map.df))%>%
  dplyr::select(mdf)%>%
  unnest()%>%
  spread(Gear.Type, layer)

# Calculating sea snake CPUE

den <- snakes_den%>%
  group_by(Species)%>%
  nest()%>%
  mutate(m = map(data, ~map.extract(df = ., var = "n", func = 'sum')),
         mdf = map(m, map.df))%>%
  dplyr::select(mdf)%>%
  unnest()%>%
  spread(Species, layer)

# Calculating mean depth in cell

library(marmap)

depth <- readGEBCO.bathy("./Data/gebco_2020_n16.7_s15.5_w72.0_e73.9.nc")

depth = fortify.bathy(depth)%>% # converting to usable data frame
  dplyr::rename(lon = x, lat = y, depth = z)%>%
  filter(depth < 1)


dep <- depth%>%
  nest()%>%
  mutate(m = map(data, ~map.extract(df = ., var = "depth", func = mean)),
         mdf = map(m, map.df))%>%
  dplyr::select(mdf)%>%
  unnest()%>%
  rename(mean.depth = layer)

# Joining creating combined data frame 

fi_den<- inner_join(fi, den, by = c("x", "y"))%>%
  inner_join(dep, by = c("x", "y"))%>%
  # calculating relative abaundance of sea snakes in each grid cell
  mutate(rel.prop = HC/(HC+HS))

# saving fi_den for analysis

write.csv(fi, "./Data/Fishing intensity_grid.csv", row.names = F)
```

## Spatial overlap between fisheries and sea snakes

```{r}

# Calculating spatial overlap between sea snakes and fisheries

sp.ovlp <- fi_den%>%
  gather(c("HS", "HC"), key = Species, value = CPUE)%>%
  gather(c("GillNet", "Trawler"), key = Gear, value = intensity)%>%
  group_by(Gear, Species)%>%
  summarise(overlap = sum(CPUE > 0 & intensity > 0, na.rm = T),
            extent.sp = sum(CPUE>0),
            rel.ovlp = overlap/extent.sp)

# Creating pretty table

sp.ovlp%>%
    dplyr::select(-extent.sp)%>%
    mutate(rel.ovlp = 100*rel.ovlp,
           Species = ifelse(Species == "HC", "Hydrophis curtus", "Hydrophis schistosus"))%>%
    rename(`Relative Overlap %` = rel.ovlp,
           `Spatial Overlap` = overlap)

```

### Testing spatial overlap between sea snakes and fisheries

We tested the spatial overlap between sea snakes and fisheries along the Sindhudurg coast as the difference in relative proportion of overlap among species and gear.

```{r}

sp.ovlp%>%
  mutate(Species = ifelse(Species == "HC", "Hydrophis curtus", "Hydrophis schistosus"))%>%
  group_by(Species)%>%
  nest()%>%
  mutate(ptest = map(data, ~prop.test(.$overlap, .$extent.sp)), # Z - test of proportion
         sumry = map(ptest, broom::tidy),
         h = map(sumry, ~pwr::ES.h(.$estimate2, .$estimate1)))%>% # effect size
  dplyr::select(sumry, h)%>%
  unnest()%>%
  dplyr::select(Species:p.value, h)%>%
  rename(`Gill Net` = estimate1,
         Trawler = estimate2)


```

Trawlers had significantly higher spatial overlap with both species. _H. curtus_ overlaped more with trawlers than _H. schistosus_ and vice - a - versa for _H. shcistosus_ in gill nets.
