==> 06-Fisheries_diet-overlap.rmd <==
# Overlap between fisheries catch and sea snake diet

```{r}
require(tidyverse)

# importing catch data

catch <- read.csv("./Data/catch.csv")

# importing gut content data

gutcontent <- read.csv("./Data/Sea_snakes_gut_content_2018-19.csv")

# Standarising sea snake gut content data

gutcontent = gutcontent%>%
  # removing specimens collected from fish landing centers
  filter(Snake.Species == "Hydrophis schistosus" | Snake.Species == "Hydrophis curtus",
         # removing unidentified specimens
         Prey.Family != "Unidentified", Prey.Family != "")%>%
  group_by(Snake.Species)%>%
  # caluclating abundance
  mutate(n = n())%>%
  group_by(Snake.Species, Prey.Family)%>%
  # calculating relative abunance from each snakes species
  summarise(Abundance = n(), n = last(n), Rel.prop = last(Abundance/n))
```

**Note**: Catch data was combined from multiple sources (Sharma et al. unpublished data and Gupta et al. unpublished data). The same was standardised for this analysis. Kindly refer to `Functions/clean catch data.R` for further details.

## Sampling Adequacy for fisherie catch data

```{r}
# calculating number trips and fishing effort

catch%>%
  group_by(Gear.Type, Sample)%>%
  # Number hauls and haul duration for each trip
  summarise(Haul.time = last(Haul.time),
            No.hauls = last(No.hauls))%>%
  group_by(Gear.Type)%>%
  # Number trips and effort sampled by gear type
  summarise(N = length(unique(Sample)),
            Haul.Hours = sum(No.hauls*Haul.time, na.rm = T)/60)

```

## Difference in catch landed per trip by gear

```{r}

# importing catch weight data

tonnage <- read.csv("./Data/catch tonnage.csv")

## Summary

tonnage%>%
  skimr::skim(Total.Catch..kg.)%>%
  skimr::yank("numeric")

# Testing assumption of normality

broom::tidy(shapiro.test(log(tonnage$Total.Catch..kg.)))

# Test

broom::tidy(t.test(log(Total.Catch..kg.) ~ Gear.Type, data = tonnage))%>%
  mutate(d = lsr::cohensD(log(Total.Catch..kg.) ~ Gear.Type, data = tonnage))%>%# effect size
  dplyr::select(estimate1:p.value, d)%>%
  # inverse log
  mutate(estimate1 = exp(estimate1),
         estimate2 = exp(estimate2))%>%
  rename(`Gill net` = estimate1,
         Trawler = estimate2)

```

Trawlers landed significantly higher catches per trip than gill nets.

**Note:** Total catch landed per trip in kgs was log transformed for normality.

## Richness of fish families found in fisheries catch

```{r}

#Richness

catch%>%
  group_by(Gear.Type)%>%
  filter(Family != "")%>%
  summarise(Family.Richness = length(unique(Family)))

```

## No. of Sea snake prey families found in fish catch and overlap

```{r}

# Creating separate prey data frame for each species

gc.hs <- filter(gutcontent, Snake.Species == "Hydrophis schistosus")

gc.hc <- filter(gutcontent, Snake.Species == "Hydrophis curtus")

# Creating data martix for fisheries catch

catch_fam <- catch%>%
  # removing unidentified speciemens
  filter(Family != "")%>%
  group_by(Gear.Type, Sample)%>%
  # Calculating total weight per sample
  mutate(Sample.Wt = sum(Weight.g, na.rm = T))%>%
  group_by(Gear.Type, Sample, Family)%>%
  # Weight of each fish family in catch
  summarise(Biomass = sum(Weight.g, na.rm = T),
            # Marking sea snake prey families
            Prey = last(ifelse(Family%in%gutcontent$Prey.Family, "Yes", "No")),
            Sample.Wt = last(Sample.Wt))%>%
  # Caluclating relative proportion in catch
  mutate(Rel.biomass = Biomass/Sample.Wt)#%>%

## Adding sea snake species

catch_fam = catch_fam%>%
  mutate(HS = if_else(Family%in%gc.hs$Prey.Family, "Yes", "No"),
         HC= if_else(Family%in%gc.hc$Prey.Family, "Yes", "No"))

# number of prey families caught by each gear

catch_fam%>%
  gather(c("HC", "HS"), key = "Snake species", value = "Prey")%>%
  filter(Prey == "Yes")%>%
  group_by(Gear.Type, `Snake species`)%>%
  summarise(`Prey Family Overlap` = length(unique(Family)))%>%
  mutate(`Snake species` = ifelse(`Snake species` == "HC", "Hydrophis curtus", "Hydrophis schistosus"))%>%
  spread(Gear.Type, `Prey Family Overlap`)

```

Trawlers caught more sea snake prey families than gillnets.

## Relative proportion of sea snake prey in fisheries catch

```{r}
# prey of each species as relative proportion in fisheries catch

catch_fam%>%
  gather(c("HC", "HS"), key = snake, value = Prey)%>%
  filter(Prey == "Yes")%>%
  group_by(Gear.Type, Sample, snake)%>%
  summarise(N = length(unique(Family)),
            Biomass = sum(Biomass),
            Sample.Wt = last(Sample.Wt),
            rel.prop = Biomass/Sample.Wt)%>%
  group_by(Gear.Type, snake)%>%
  summarise(Mean.prop = mean(rel.prop))%>%
  spread(Gear.Type, Mean.prop)
  
```

Sea snake prey consisted of a greater proportion of gillnet catch than trawler catch on average.

## Is the proportion of _H. curtus_ prey greater than that of _H. schistosus_ in catch?

```{r}

catch_fam%>%
  gather(c("HC", "HS"), key = snake, value = Prey)%>%
  filter(Prey == "Yes")%>%
  group_by(Gear.Type, Sample, snake)%>%
  summarise(Biomass = sum(Biomass),
            Sample.Wt = last(Sample.Wt),
            rel.prop = Biomass/Sample.Wt)%>%
  group_by(Gear.Type)%>%
  nest()%>%
  mutate(ttest = map(data, ~t.test(rel.prop ~ snake, data = .)),
         sumry = map(ttest, broom::tidy),
         d = map(data, ~lsr::cohensD(rel.prop ~ snake, data = .)))%>%
  dplyr::select(sumry, d)%>%
  unnest()%>%
  dplyr::select(estimate1:parameter, d, -estimate)%>%
  rename(`H. curtus` = estimate1,
         `H. schistosus` = estimate2)

```

Yes, for trawlers.


## Sea snake prey species in fisheries catch

Species constituting >10% of the catch on average are represented.

```{r fig.height=6}

catch_fam%>%
  gather(c("HC", "HS"), key = "snake", value = "Prey")%>%
  filter(Prey == "Yes")%>%
  mutate(snake = ifelse(snake == "HC", "Hydrophis curtus",  "Hydrophis schistosus"),
         Gear.Type = ifelse(Gear.Type == "Trawler", "Trawl Net",  Gear.Type))%>%
  group_by(Gear.Type, snake)%>%
  summarise(N = length(unique(Family)),
            p = mean(Rel.biomass, na.rm = T),
            sd = sd(Rel.biomass, na.rm = T)/sqrt(n()))%>%
  ggplot(aes(snake, p, fill = Gear.Type))+
  geom_col(width = 0.5, position = position_dodge(), col = "black")+
  geom_errorbar(aes(ymin = p-sd, ymax = p+sd), position = position_dodge(width = .5), width = 0.25)+
  scale_fill_brewer(palette = "Greys", name = "Gear Type")+
  scale_y_continuous(expand = c(0,0), limits = c(0,0.3))+
  labs(x = "Snake Species", y = "Average proportion \n in catch")+
  theme(axis.text.x = element_text(face = "italic"))

ggsave(last_plot(), filename = "./Figures and Tables/figure5.png", height = 8, width = 8)
```
